[Chore] browser required: move `recentOpenService.ts` into `browser/` dir
